abstract out renderMultinomialConfusionMatrix fix #105
